addresses_deployers.sol - uint => uint256

+++ src/test/addresses_deployers.sol
@@ -32,7 +32,7 @@ contract Deployers {
         ];
     }
 
-    function count() external view returns (uint) {
+    function count() external view returns (uint256) {
         return addr.length;
     }
 }
